% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction.R
\name{AddStarvationToPreyReproductiveRate}
\alias{AddStarvationToPreyReproductiveRate}
\title{Add starvation to prey reproductive rate}
\usage{
AddStarvationToPreyReproductiveRate(PctDefol, Constant, PreyReproRate)
}
\arguments{
\item{PctDefol}{DESCRIPTION NEEEDED}

\item{Constant}{DESCRIPTION NEEEDED}

\item{PreyReproRate}{DESCRIPTION NEEEDED}
}
\value{
DESCRIPTION NEEEDED
}
\description{
Add starvation to prey reproductive rate
}
