% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predation.R
\name{ApplyDensityIndependentLoss}
\alias{ApplyDensityIndependentLoss}
\title{Apply density-independent loss}
\usage{
ApplyDensityIndependentLoss(PreyCountRaster)
}
\arguments{
\item{PreyCountRaster}{\code{RasterLayer}. DESCRIPTION NEEDED.}
}
\value{
egg count DESCRIPTION NEEDED
}
\description{
Calculate density independent predation.
Apply a high scale spatial autocorrelated function to the budworm count to simulate
density-independent predation.
This simulates the effect of summer temperature (huge scale spatial autocorrelation) on ability
to escape density-independent predation from thermoregulatory predators.
}
