% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{CalcLDDRatio}
\alias{CalcLDDRatio}
\title{Calculate long-distance dispersal ratio}
\usage{
CalcLDDRatio(
  PctDefol,
  MinLDD,
  HalfLDD,
  MaxLDD,
  MaxLDDProp,
  PositiveRelation,
  Constant
)
}
\arguments{
\item{PctDefol}{TODO}

\item{MinLDD}{TODO}

\item{HalfLDD}{TODO}

\item{MaxLDD}{TODO}

\item{MaxLDDProp}{TODO}

\item{PositiveRelation}{TODO}

\item{Constant}{TODO}
}
\description{
Calculate long-distance dispersal ratio
}
