% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction.R
\name{CalculatePredatorReproductiveRate}
\alias{CalculatePredatorReproductiveRate}
\title{Calculate predator reproductive rate}
\usage{
CalculatePredatorReproductiveRate(
  ParamB,
  ParamC,
  BackgrPred,
  PreyDensity,
  PredatorDensity,
  maxReproEnemy
)
}
\arguments{
\item{ParamB}{DESCRIPTION NEEEDED}

\item{ParamC}{DESCRIPTION NEEEDED}

\item{BackgrPred}{DESCRIPTION NEEEDED}

\item{PreyDensity}{DESCRIPTION NEEEDED}

\item{PredatorDensity}{DESCRIPTION NEEEDED}

\item{maxReproEnemy}{DESCRIPTION NEEEDED}
}
\value{
DESCRIPTION NEEEDED
}
\description{
Calculate predator reproductive rate
}
