% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction.R
\name{CalculatePreyReproductiveRate}
\alias{CalculatePreyReproductiveRate}
\title{Calculate prey reproductive rate}
\usage{
CalculatePreyReproductiveRate(
  ParamA,
  ParamB,
  ParamC,
  Interference,
  PreyDensity,
  PredatorDensity,
  MatingEff,
  maxReproPrey,
  Fecundity
)
}
\arguments{
\item{ParamA}{DESCRIPTION NEEEDED}

\item{ParamB}{DESCRIPTION NEEEDED}

\item{ParamC}{DESCRIPTION NEEEDED}

\item{Interference}{DESCRIPTION NEEEDED}

\item{PreyDensity}{DESCRIPTION NEEEDED}

\item{PredatorDensity}{DESCRIPTION NEEEDED}

\item{MatingEff}{DESCRIPTION NEEEDED}

\item{maxReproPrey}{DESCRIPTION NEEEDED}

\item{Fecundity}{DESCRIPTION NEEEDED}
}
\value{
DESCRIPTION NEEEDED
}
\description{
Calculate prey reproductive rate
}
