% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{CalcWinterMortalityAndDispersal}
\alias{CalcWinterMortalityAndDispersal}
\title{CalcWinterMortalityAndDispersal}
\usage{
CalcWinterMortalityAndDispersal(
  WindowDistance,
  Count,
  SurvivingPercent,
  LDisperse
)
}
\arguments{
\item{WindowDistance}{DESCRIPTION NEEDED}

\item{Count}{DESCRIPTION NEEDED}

\item{SurvivingPercent}{DESCRIPTION NEEDED}

\item{LDisperse}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Calculate overwintering loss and do a focal mean to simulate L2 dispersal.
Focal mean is done with equal weight of the 8 neighbor cells by default.
}
\author{
Louis-Etienne Robert
}
