% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predation.R
\name{ConvertToEnemyDensity}
\alias{ConvertToEnemyDensity}
\title{Convert to enemy density}
\usage{
ConvertToEnemyDensity(EnemyCount, InsectCount)
}
\arguments{
\item{EnemyCount}{DESCRIPTION NEEDED}

\item{InsectCount}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Convert to density using the current budworm count (before scaling the budworm count).
}
