% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasters.R
\name{createRaster}
\alias{createRaster}
\title{Create dummy rasters based on raster to match}
\usage{
createRaster(
  rtm = NULL,
  gaus = TRUE,
  rownumb = 20,
  colnumb = 20,
  xmn = -10,
  xmx = 10,
  ymn = -10,
  ymx = 10,
  sc = 100,
  vr = 5
)
}
\arguments{
\item{rtm}{\code{RasterLayer} to use as template.}

\item{gaus}{logical. If \code{TRUE} (default), creates raster using Gaussian map.
If \code{FALSE}, creates a raster with all pixel values equal to \code{1}.}

\item{rownumb}{number of rows for the new raster, if \code{rtm} not provided.}

\item{colnumb}{number of columns for the new raster, if \code{rtm} not provided.}

\item{xmn}{min x extent, if \code{rtm} not provided.}

\item{xmx}{max x extent, if \code{rtm} not provided.}

\item{ymn}{min y extent, if \code{rtm} not provided.}

\item{ymx}{max y extent, if \code{rtm} not provided.}

\item{sc}{spatial scale, passed to \code{gaussMap}.}

\item{vr}{spatial variance, passed to \code{gaussMap}.}
}
\value{
\code{RasterLayer}
}
\description{
Create dummy rasters based on raster to match
}
\author{
Louis-Etienne Robert and Alex Chubaty
}
