% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defoliation.R
\name{DeterminePctDefoliation}
\alias{DeterminePctDefoliation}
\title{Determine percent defoliation}
\usage{
DeterminePctDefoliation(
  ParamA,
  ParamB,
  ParamC,
  Interference,
  PreyDensity,
  PredatorDensity,
  LarvaeConsump,
  lambda
)
}
\arguments{
\item{ParamA}{DESCRIPTION NEEDED}

\item{ParamB}{DESCRIPTION NEEDED}

\item{ParamC}{DESCRIPTION NEEDED}

\item{Interference}{DESCRIPTION NEEDED}

\item{PreyDensity}{DESCRIPTION NEEDED}

\item{PredatorDensity}{DESCRIPTION NEEDED}

\item{LarvaeConsump}{DESCRIPTION NEEDED}

\item{lambda}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Determine percent defoliation
}
