% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{GetLDDFlight}
\alias{GetLDDFlight}
\title{\code{GetLDDFlight}}
\usage{
GetLDDFlight(rZY, MaxLDDProp, PosRel)
}
\arguments{
\item{rZY}{TODO}

\item{MaxLDDProp}{TODO}

\item{PosRel}{TODO}
}
\value{
TODO
}
\description{
Default values of Max,Half,Min bind the data between 0 and 1 and use the defol value as the lddflight value
}
