% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{GetLDDHabitat}
\alias{GetLDDHabitat}
\title{GetLDDHabitat}
\usage{
GetLDDHabitat(defol, MinLDD, HalfLDD, MaxLDD)
}
\arguments{
\item{defol}{DESCRIPTION NEEDED}

\item{MinLDD}{DESCRIPTION NEEDED}

\item{HalfLDD}{DESCRIPTION NEEDED}

\item{MaxLDD}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
GetLDDHabitat
}
