% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defoliation.R
\name{GetMortalityRate}
\alias{GetMortalityRate}
\title{Get mortality rate}
\usage{
GetMortalityRate(species, age, cumDefol)
}
\arguments{
\item{species}{DESCRIPTION NEEDED}

\item{age}{DESCRIPTION NEEDED}

\item{cumDefol}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Get mortality rate
}
