% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{LocalDispersalEnemy}
\alias{LocalDispersalEnemy}
\title{LocalDispersalEnemy}
\usage{
LocalDispersalEnemy(
  EnemyCount,
  InsectCount,
  WindowDistanceEnemy,
  Proportion,
  PreyBiasProportion
)
}
\arguments{
\item{EnemyCount}{\code{RasterLayer}. DESCRIPTION NEEDED}

\item{InsectCount}{\code{RasterLayer}. DESCRIPTION NEEDED}

\item{WindowDistanceEnemy}{DESCRIPTION NEEDED}

\item{Proportion}{DESCRIPTION NEEDED}

\item{PreyBiasProportion}{DESCRIPTION NEEDED}
}
\value{
DESCRIPTION NEEDED
}
\description{
Calculate overwintering loss and do a focal mean to simulate local dispersal.
Focal mean is done with equal weight of the 8 neighbour cells.
}
