% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{meanNZ}
\alias{meanNZ}
\title{Calculate the mean of non-zero values}
\usage{
meanNZ(x)
}
\arguments{
\item{x}{DESCRIPTION NEEEDED}
}
\value{
numeric mean value
}
\description{
Used as input for focal to handle edges padded with 0, and 0 values.
}
\author{
Louis-Etienne Robert
}
